+++ CesiumGltf/src/AccessorUtility.cpp
@@ -3,6 +3,65 @@
 #include "CesiumGltf/Model.h"
 
 namespace CesiumGltf {
+FeatureIdAccessorType GetFeatureIdAccessorView(
+    const Model& model,
+    const MeshPrimitive& primitive,
+    int32_t featureIdAttributeIndex) {
+  const std::string attributeName =
+      "_FEATURE_ID_" + std::to_string(featureIdAttributeIndex);
+  auto featureId = primitive.attributes.find(attributeName);
+  if (featureId == primitive.attributes.end()) {
+    return FeatureIdAccessorType();
+  }
+
+  const Accessor* pAccessor =
+      model.getSafe<Accessor>(&model.accessors, featureId->second);
+  if (!pAccessor || pAccessor->type != Accessor::Type::SCALAR ||
+      pAccessor->normalized) {
+    return FeatureIdAccessorType();
+  }
+
+  switch (pAccessor->componentType) {
+  case Accessor::ComponentType::BYTE:
+    return AccessorView<int8_t>(model, *pAccessor);
+  case Accessor::ComponentType::UNSIGNED_BYTE:
+    return AccessorView<uint8_t>(model, *pAccessor);
+  case Accessor::ComponentType::SHORT:
+    return AccessorView<int16_t>(model, *pAccessor);
+  case Accessor::ComponentType::UNSIGNED_SHORT:
+    return AccessorView<uint16_t>(model, *pAccessor);
+  case Accessor::ComponentType::FLOAT:
+    return AccessorView<float>(model, *pAccessor);
+  default:
+    return FeatureIdAccessorType();
+  }
+}
+
+IndexAccessorType
+GetIndexAccessorView(const Model& model, const MeshPrimitive& primitive) {
+  if (primitive.indices < 0) {
+    return IndexAccessorType();
+  }
+
+  const Accessor* pAccessor =
+      model.getSafe<Accessor>(&model.accessors, primitive.indices);
+  if (!pAccessor || pAccessor->type != Accessor::Type::SCALAR ||
+      pAccessor->normalized) {
+    return AccessorView<uint8_t>();
+  }
+
+  switch (pAccessor->componentType) {
+  case Accessor::ComponentType::UNSIGNED_BYTE:
+    return AccessorView<uint8_t>(model, *pAccessor);
+  case Accessor::ComponentType::UNSIGNED_SHORT:
+    return AccessorView<uint16_t>(model, *pAccessor);
+  case Accessor::ComponentType::UNSIGNED_INT:
+    return AccessorView<uint32_t>(model, *pAccessor);
+  default:
+    return AccessorView<uint8_t>();
+  }
+}
+
 TexCoordAccessorType GetTexCoordAccessorView(
     const Model& model,
     const MeshPrimitive& primitive,
@@ -14,9 +73,8 @@ TexCoordAccessorType GetTexCoordAccessorView(
     return TexCoordAccessorType();
   }
 
-  const Accessor* pAccessor = model.getSafe<Accessor>(
-      &model.accessors,
-      primitive.attributes.at(texCoordName));
+  const Accessor* pAccessor =
+      model.getSafe<Accessor>(&model.accessors, texCoord->second);
   if (!pAccessor || pAccessor->type != Accessor::Type::VEC2) {
     return TexCoordAccessorType();
   }